+++ gov.usgs.warc.iridium.sbd.decoder/src/main/java/gov/usgs/warc/iridium/sbd/decoder/parser/PseudobinaryBPayloadDecoder.java
@@ -1,10 +1,9 @@
 package gov.usgs.warc.iridium.sbd.decoder.parser;
 
 import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkElementIndex;
 
-import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
+import com.google.common.collect.Queues;
 import gov.usgs.warc.iridium.sbd.decoder.parser.elements.Payload;
 import gov.usgs.warc.iridium.sbd.decoder.parser.elements.PayloadType;
 import gov.usgs.warc.iridium.sbd.decoder.sixbitbinary.Decode;
@@ -13,7 +12,12 @@ import gov.usgs.warc.iridium.sbd.domain.SbdDecodeOrder;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.Queue;
 import java.util.SortedSet;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Decoder for {@link PayloadType#PSEUDOBINARY_B_DATA_FORMAT}
@@ -24,6 +28,13 @@ import java.util.SortedSet;
  */
 public class PseudobinaryBPayloadDecoder implements PayloadDecoder
 {
+
+	/**
+	 * @author mckelvym
+	 * @since Mar 26, 2019
+	 */
+	private static final Logger log = LoggerFactory
+			.getLogger(PseudobinaryBPayloadDecoder.class);
 
 	/**
 	 * @author mckelvym
@@ -46,13 +57,16 @@ public class PseudobinaryBPayloadDecoder implements PayloadDecoder
 				.getPayloadType() == PayloadType.PSEUDOBINARY_B_DATA_FORMAT,
 				"Invalid payload type for this decoder.");
 
-		final List<Byte> payloadBytes = Lists.newArrayList();
-		for (final byte b : p_Payload.getPayload())
-		{
-			payloadBytes.add(b);
-		}
+		final byte[] payload = p_Payload.getPayload();
+		log.info(String.format("Payload:\n%s", new String(payload)));
+
+		final Queue<Byte> payloadQueue = Queues
+				.newArrayBlockingQueue(payload.length);
+		IntStream.range(0, payload.length)
+				.forEach(i -> payloadQueue.offer(payload[i]));
 
 		final Map<SbdDataType, Double> dataMap = Maps.newLinkedHashMap();
+
 		/**
 		 * Build the map of data type and its corresponding value decoded from
 		 * the payload bytes.
@@ -61,18 +75,38 @@ public class PseudobinaryBPayloadDecoder implements PayloadDecoder
 		{
 			final SbdDataType datatype = order.getDatatype();
 			final int byteLength = datatype.getBytes();
-			final int startIndex = (int) order.getByteOffset();
 
-			checkElementIndex(startIndex, payloadBytes.size(), String.format(
-					"The payload (%s; size: %s) does not have enough bytes to decode '%s' starting at byte %s.",
-					Arrays.toString(payloadBytes.toArray()),
-					payloadBytes.size(), datatype, startIndex));
+			if (payloadQueue.size() >= byteLength)
+			{
+				final List<Byte> decodeList = IntStream.range(0, byteLength)
+						.mapToObj(i -> payloadQueue.poll())
+						.collect(Collectors.toList());
+				float value = Float.NaN;
+				try
+				{
+					value = Decode.valueAtIndex(decodeList, 0,
+							decodeList.size(), 1);
+				}
+				catch (final IllegalArgumentException e)
+				{
+					log.warn(String.format(
+							"Unable to decode '%s' (%s): %s%nUsing NaN instead.",
+							datatype, Arrays.toString(decodeList.toArray()),
+							e.getMessage()));
+				}
+				final double transformedVal = Float.isNaN(value) ? Float.NaN
+						: datatype.transformValue(value);
 
-			final float value = Decode.valueAtIndex(payloadBytes, startIndex,
-					byteLength, 1);
-			final double transformedVal = datatype.transformValue(value);
-
-			dataMap.put(datatype, transformedVal);
+				log.info(String.format("Decoded '%s': %s -> %s", datatype,
+						value, transformedVal));
+				dataMap.put(datatype, transformedVal);
+			}
+			else
+			{
+				log.warn(String.format(
+						"The remaining payload (size: %s) does not have enough bytes to decode '%s': need %s byte(s).",
+						payloadQueue.size(), datatype, byteLength));
+			}
 		}
 		return dataMap;
 	}